+++ src/dynnn/encoding.py
